requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<12.0
rembg>=2.0,<3.0
